refactor: CoreData 관련 코드를 DataModel 클래스로 생성
